refactor(tape): split application services
